New version of evolution2.py. Now updates separate genepool populations in the 'current' folder. References to evolution.py have not been updated yet, so this implementation is not yet working.
First, merge with the most up-to-date version of the rest of the code. Merge conflicts expected in presets.py, as the initialization of the 'current' folder has changed.

Second, change the most up-to-date version to reference the new pool of files and draw genes from these files. This might require extensive recoding of geometry.py, genetics.py and possibly others...

**Changed some files to enable merge into master**